+++ .gitignore
@@ -291,3 +291,42 @@ marimo/\_lsp/
 # Streamlit
 
 .streamlit/secrets.toml
+
+# Node modules
+node_modules/
+
+# Compiled JavaScript and Map files
+*.js
+*.js.map
+dist/
+build/
+
+# Log files
+*.log
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
+
+# Editor-specific files
+.vscode/ # VSCode settings
+.idea/ # WebStorm/IntelliJ settings
+*.sublime-project
+*.sublime-workspace
+
+# Operating System files
+.DS_Store
+Thumbs.db
+
+# Environment variables
+.env
+.env.local
+
+# npm/yarn cache
+.npm/
+.yarn/cache/
+
+# Test coverage
+coverage/
+
+# Other temporary files
+tmp/
