+++ src/schema-extractor/ts/.prettierignore
