+++ src/schema-extractor/ts/src/cli.ts
