+++ src/schema-extractor/ts/tests/fixtures/cloesce-config.json
@@ -1,0 +1,3 @@
+{
+    "source": "."
+}
