+++ src/schema-extractor/ts/tsconfig.json
