+++ src/schema-extractor/ts/tsconfig.test.json
